---python/requirements.txt
ipykernel
requests
dnspython
pandas
streamlit